add ingredients route and separated component from SearchByIngredient
