+++ src/router/index.ts
@@ -34,7 +34,7 @@ const routes: RouteRecordRaw[] = [
             component: Ingredients,
         },
         {
-            path: '/by-ingredient/:ingredient?',
+            path: '/by-ingredient/:ingredient',
             name: 'byIngredient',
             component: SearchByIngredient,
         },
