+++ src/views/SearchByIngredient.vue
@@ -1,32 +1,22 @@
 <script setup lang="ts">
-import { onMounted, ref } from 'vue';
-import axiosClient from '../utils/axiosClient'
-    
-const ingredients = ref({});
+import { computed, onMounted } from 'vue';
+import store from '../store';
+import { useRoute } from 'vue-router';
+
+
+const route = useRoute();
+const meals = computed(() => store.state.mealsByIngredient)
 
 onMounted(() => {
-    axiosClient.get(`list.php?i=list`)
-        .then(({data}) => {
-            ingredients.value = data.meals;
-        })
-});
+    store.dispatch('searchMealsByIngredient', route.params.ingredient)
+})
 </script>
 
 <template>
-    <div>
-        <!-- <pre>{{ ingredients }}</pre> -->
-        <router-link 
-            v-for="ingredient of ingredients"
-            :key="ingredient.idIngredient"
-            :to="{
-                name: 'byIngredient',
-                params: { ingredient: ingredient.idIngredient },
-            }"
-        >
-            <h3>{{ ingredient.strIngredient }}</h3>
-            <div>{{ ingredient.strDescription }}</div>
-            <br/>
-            <hr />
-        </router-link>
+    <div class="grid grid-cols-1 md:grid-cols-3 xl:grid-cols-4 gap-5">
+        <MealItem v-for="meal of meals" :key="meal.idMeal" :meal="meal" />
+    </div>
+    <div v-if="!meals.length">
+        There are no meals
     </div>
 </template>
